backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "VibeCAD backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.118.0",
    "google-genai>=1.41.0",
    "scad2gltf>=0.0.3",
]
